// src/api/models/Stream.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jingle Jam Community Page Public API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Creator } from './Creator';
import {
    CreatorFromJSON,
    CreatorFromJSONTyped,
    CreatorToJSON,
} from './Creator';
import type { StreamVodsInner } from './StreamVodsInner';
import {
    StreamVodsInnerFromJSON,
    StreamVodsInnerFromJSONTyped,
    StreamVodsInnerToJSON,
} from './StreamVodsInner';

/**
 * 
 * @export
 * @interface Stream
 */
export interface Stream {
    /**
     * 
     * @type {string}
     * @memberof Stream
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof Stream
     */
    subtitle?: string;
    /**
     * 
     * @type {string}
     * @memberof Stream
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof Stream
     */
    markdownDescription?: string;
    /**
     * 
     * @type {Date}
     * @memberof Stream
     */
    start: Date;
    /**
     * 
     * @type {Date}
     * @memberof Stream
     */
    end: Date;
    /**
     * 
     * @type {Array<Creator>}
     * @memberof Stream
     */
    creators?: Array<Creator>;
    /**
     * 
     * @type {Array<StreamVodsInner>}
     * @memberof Stream
     */
    vods?: Array<StreamVodsInner>;
    /**
     * 
     * @type {string}
     * @memberof Stream
     */
    color: string;
}

/**
 * Check if a given object implements the Stream interface.
 */
export function instanceOfStream(value: object): value is Stream {
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('start' in value) || value['start'] === undefined) return false;
    if (!('end' in value) || value['end'] === undefined) return false;
    if (!('color' in value) || value['color'] === undefined) return false;
    return true;
}

export function StreamFromJSON(json: any): Stream {
    return StreamFromJSONTyped(json, false);
}

export function StreamFromJSONTyped(json: any, ignoreDiscriminator: boolean): Stream {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'],
        'subtitle': json['subtitle'] == null ? undefined : json['subtitle'],
        'description': json['description'] == null ? undefined : json['description'],
        'markdownDescription': json['markdownDescription'] == null ? undefined : json['markdownDescription'],
        'start': (new Date(json['start'])),
        'end': (new Date(json['end'])),
        'creators': json['creators'] == null ? undefined : ((json['creators'] as Array<any>).map(CreatorFromJSON)),
        'vods': json['vods'] == null ? undefined : ((json['vods'] as Array<any>).map(StreamVodsInnerFromJSON)),
        'color': json['color'],
    };
}

export function StreamToJSON(value?: Stream | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'title': value['title'],
        'subtitle': value['subtitle'],
        'description': value['description'],
        'markdownDescription': value['markdownDescription'],
        'start': ((value['start']).toISOString()),
        'end': ((value['end']).toISOString()),
        'creators': value['creators'] == null ? undefined : ((value['creators'] as Array<any>).map(CreatorToJSON)),
        'vods': value['vods'] == null ? undefined : ((value['vods'] as Array<any>).map(StreamVodsInnerToJSON)),
        'color': value['color'],
    };
}